app/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
av==16.0.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
coloredlogs==15.0.1
ctranslate2==4.6.0
distro==1.9.0
edge-tts==6.1.5
faster-whisper==1.2.1
filelock==3.20.0
flatbuffers==25.9.23
pystray==0.19.5
qrcode==7.4.2
frozenlist==1.8.0
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.0.1
humanfriendly==10.0
idna==3.11
jiter==0.11.1
keyboard==0.13.5
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.7.0
numpy==2.3.4
ollama==0.6.0
onnxruntime==1.23.2
openai==2.6.1
packaging==25.0
pefile==2023.2.7
pillow==12.0.0
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.3
pyaudio==0.2.14
pycaw==20240210
comtypes==1.4.6
PyAutoGUI==0.9.54
py-cord==2.6.0
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pygame==2.6.1
PyGetWindow==0.0.9
pyinstaller-hooks-contrib==2025.9
PyMsgBox==2.0.1
PyNaCl==1.5.0
pyperclip==1.11.0
pyqt5==5.15.11
pyreadline3==3.5.4
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
python-dotenv==1.2.1
pytweening==1.2.0
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.3
redis==7.0.1
requests==2.32.5
screeninfo==0.8.1
shellingham==1.5.4
sniffio==1.3.1
sounddevice==0.5.3
soundfile==0.13.1
spotipy==2.25.1
sympy==1.14.0
tabulate==0.9.0
tokenizers==0.22.1
tqdm==4.67.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
yarl==1.22.0
python-multipart==0.0.9
